[구현]공격 상호작용, 네트워크 통신 기본 기능 구현 종료, [수정]ui갱신 이상 수정

+++ LeeGameServer/ConsoleApplication1/Packets.cpp
@@ -110,8 +110,8 @@ void WinPlayerIdPacket::Deserialize(Json::Value& root) {
 void MapDataPacket::Serialize(Json::Value& root) {
 	root["type"] = (int)type;
 	for (auto a : furniturePos) {
-		root["furniturePosX"].append(a.x);
-		root["furniturePosY"].append(a.y);
+		root["furniturePosX"].append((int)a.x);
+		root["furniturePosY"].append((int)a.y);
 	}
 }
